requirements.txt
python-telegram-bot==11.1.0
google-cloud-dialogflow==2.1.2
vk-api==11.9.4
python-dotenv==0.17.1